Added more assets and increased resolution on boss arena and floor panel
I have increased the resolution to the boss arena and the floor panel, I have made changed to the assets I have incorporated some foliage for small meteorites, can be reduced if needs be. I have attached a ice material to the barrier to give a visual interpretation to a barrier being there, this can be changed at a later date.